+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # development version
 
+* New `cli_memo()` function to create a list of items or tasks.
+
 * New `cli::cli()` function to create a cli message from multiple
   cli calls (#170).